fix int32 to int in build dstore

+++ build_dstore.py
@@ -27,7 +27,7 @@ if args.dstore_fp16:
     vals = np.memmap(args.dstore_mmap+'_vals.npy', dtype=np.int16, mode='r', shape=(args.dstore_size, 1))
 else:
     keys = np.memmap(args.dstore_mmap+'_keys.npy', dtype=np.float32, mode='r', shape=(args.dstore_size, args.dimension))
-    vals = np.memmap(args.dstore_mmap+'_vals.npy', dtype=np.int32, mode='r', shape=(args.dstore_size, 1))
+    vals = np.memmap(args.dstore_mmap+'_vals.npy', dtype=np.int, mode='r', shape=(args.dstore_size, 1))
 
 if not os.path.exists(args.faiss_index+".trained"):
     # Initialize faiss index
